Now using the byte struct in types to access the low and high bytes of a 16 bit number. Also sending a velocity of 0 after a drive command to tell the irobot to stop moving.

+++ Drive.h
@@ -11,7 +11,7 @@
 
 #include "types.h"
 
-void driveStraight(int vel, int dist);
+void driveStraight(int16_t vel, int dist);
         
 // include shiz l8r
